scripts/requirements_api.txt
# Web API Server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
pydantic==2.5.0

# Kafka client
kafka-python==2.0.2

# Spark & Delta Lake (same as main requirements.txt)
pyspark==3.5.0
delta-spark==3.0.0

# Additional utilities
python-multipart==0.0.6